Data.Extensible.Class

Description

Class

class (Functor f, Profunctor p) => Extensible f p t where

This class allows us to use pieceAt for both sums and products.

Minimal complete definition

pieceAt

Methods

pieceAt :: Membership xs x -> Optic' p f (t h xs) (h x)

piece :: (x ∈ xs, Extensible f p t) => Optic' p f (t h xs) (h x)

Accessor for an element.

pieceAssoc :: (Associate k v xs, Extensible f p t) => Optic' p f (t h xs) (h (k :> v))

Like piece, but reckon membership from its key.

itemAt :: (Wrapper h, Extensible f p t) => Membership xs x -> Optic' p f (t h xs) (Repr h x)

Access a specified element through a wrapper.

item :: (Wrapper h, Extensible f p t, x ∈ xs) => proxy x -> Optic' p f (t h xs) (Repr h x)

Access an element through a wrapper.

itemAssoc :: (Wrapper h, Extensible f p t, Associate k v xs) => proxy k -> Optic' p f (t h xs) (Repr h (k :> v))

Access an element specified by the key type through a wrapper.

Membership

data Membership xs x

The position of x in the type level set xs.

mkMembership :: Int -> Q Exp

Generates a Membership that corresponds to the given ordinal (0-origin).

compareMembership :: Membership xs x -> Membership xs y -> Either Ordering (x :~: y)

Compare two Memberships.

leadership :: Membership (y ': xs) x -> ((x :~: y) -> r) -> (Membership xs x -> r) -> r

Embodies a type equivalence to ensure that the Membership points the first element.

Member

class Member xs x where

x is a member of xs

Minimal complete definition

membership

Methods

membership :: Membership xs x

remember :: forall xs x r. Membership xs x -> (Member xs x => r) -> r

Remember that Member xs x from Membership.

type (∈) x xs = Member xs x

Unicode flipped alias for Member

type family FindType (x :: k) (xs :: [k]) :: [Nat] where ...

FindType types

Equations

FindType x (x ': xs) = Zero ': FindType x xs

FindType x (y ': ys) = MapSucc (FindType x ys)

FindType x '[] = '[]

Generation

class Generate xs where

Every type-level list is an instance of Generate.

Minimal complete definition

henumerate, hcount, hgenerateList

Methods

henumerate :: (forall x. Membership xs x -> r -> r) -> r -> r

Enumerate all possible Memberships of xs.

hcount :: proxy xs -> Int

Count the number of memberships.

hgenerateList :: Applicative f => (forall x. Membership xs x -> f (h x)) -> f (HList h xs)

Enumerate Memberships and construct an HList.

class (ForallF c xs, Generate xs) => Forall c xs where

Every element in xs satisfies c

Minimal complete definition

henumerateFor, hgenerateListFor

Methods

henumerateFor :: proxy c -> proxy' xs -> (forall x. c x => Membership xs x -> r -> r) -> r -> r

Enumerate all possible Memberships of xs with an additional context.

hgenerateListFor :: Applicative f => proxy c -> (forall x. c x => Membership xs x -> f (h x)) -> f (HList h xs)

type family ForallF (c :: k -> Constraint) (xs :: [k]) :: Constraint where ...

Equations

ForallF c '[] = ()

ForallF c (x ': xs) = (c x, Forall c xs)

Association

data Assoc k v

The kind of key-value pairs

Constructors

k :> v infix 0

type (>:) = (:>)

A synonym for (:>)

class Associate k v xs | k xs -> v where

Associate k v xs is essentially identical to (k :> v) ∈ xs , but the type v is inferred from k and xs.

Minimal complete definition

association

Methods

association :: Membership xs (k :> v)

type family FindAssoc (key :: k) (xs :: [Assoc k v]) where ...

Equations

FindAssoc k ((k :> v) ': xs) = (Zero :> v) ': MapSuccKey (FindAssoc k xs)

FindAssoc k ((k' :> v) ': xs) = MapSuccKey (FindAssoc k xs)

FindAssoc k '[] = '[]

Sugar

type family Elaborate (key :: k) (xs :: [v]) :: Elaborated k v where ...

Equations

Elaborate k '[] = Missing k

Elaborate k '[x] = Expecting x

Elaborate k xs = Duplicate k

data Elaborated k v

A readable type search result

Constructors

Expecting v

Missing k

Duplicate k