Data.Extensible.Effect

Description

Name-based extensible effects

Base

data Instruction xs a where

A unit of named effects. This is a variant of (:|) specialised for 'Type -> Type'.

Constructors

Instruction :: !(Membership xs kv) -> AssocValue kv a -> Instruction xs a

type Eff xs = Skeleton (Instruction xs)

The extensible operational monad

liftEff :: forall s t xs a. Associate s t xs => Proxy s -> t a -> Eff xs a

Lift an instruction onto an Eff action.

liftsEff :: forall s t xs a r. Associate s t xs => Proxy s -> t a -> (a -> r) -> Eff xs r

Lift an instruction onto an Eff action and apply a function to the result.

hoistEff :: forall s t xs a. Associate s t xs => Proxy s -> (forall x. t x -> t x) -> Eff xs a -> Eff xs a

Censor a specific type of effects in an action.

castEff :: IncludeAssoc ys xs => Eff xs a -> Eff ys a

Upcast an action.

Step-wise handling

newtype Interpreter f g

Transformation between effects

Constructors

Interpreter

Fields

runInterpreter :: forall a. g a -> f a

handleEff :: RecordOf (Interpreter m) xs -> Eff xs a -> MonadView m (Eff xs) a

Process an Eff action using a record of Interpreters.

Peeling

peelEff

Arguments

:: Rebinder xs r
Re-bind an unrelated action

-> (a -> r)
return the result

-> (forall x. t x -> (x -> r) -> r)
Handle the foremost type of an action

-> Eff ((k >: t) ': xs) a

-> r

Build a relay-style handler from a triple of functions.

runStateEff = peelEff1 (a s -> return (a, s))
  (m k s -> let (a, s') = runState m s in k a s')

type Rebinder xs r = forall x. Instruction xs x -> (x -> r) -> r

A function to bind an Instruction in peelEff.

rebindEff0 :: Rebinder xs (Eff xs r)

A common value for the second argument of peelEff. Binds an instruction directly.

peelEff0

Arguments

:: (a -> Eff xs r)
return the result

-> (forall x. t x -> (x -> Eff xs r) -> Eff xs r)
Handle the foremost type of an action

-> Eff ((k >: t) ': xs) a

-> Eff xs r

peelEff specialised for continuations with no argument

rebindEff1 :: Rebinder xs (a -> Eff xs r)

A pre-defined value for the second argument of peelEff. Preserves the argument of the continuation.

peelEff1

Arguments

:: (a -> b -> Eff xs r)
return the result

-> (forall x. t x -> (x -> b -> Eff xs r) -> b -> Eff xs r)
Handle the foremost type of an action

-> Eff ((k >: t) ': xs) a

-> b

-> Eff xs r

peelEff specialised for 1-argument continuation

rebindEff2 :: Rebinder xs (a -> b -> Eff xs r)

A pre-defined value for the second argument of peelEff. Preserves two arguments of the continuation.

leaveEff :: Eff '[] a -> a

Reveal the final result of Eff.

retractEff :: forall k m a. Monad m => Eff '[k >: m] a -> m a

Tear down an action using the Monad instance of the instruction.

Anonymous actions

data Action args a r where

Anonymous representation of instructions.

Constructors

AResult :: Action '[] a a

AArgument :: x -> Action xs a r -> Action (x ': xs) a r

type family Function args r :: * where ...

Function [a, b, c] r is a -> b -> c -> r

Equations

Function '[] r = r

Function (x ': xs) r = x -> Function xs r

runAction :: Function xs (f a) -> Action xs a r -> f r

Pass the arguments of Action to the supplied function.

(@!?) :: FieldName k -> Function xs (f a) -> Field (Interpreter f) (k :> Action xs a) infix 1

Create a Field of a Interpreter for an Action.

peelAction

Arguments

:: (forall x. Instruction xs x -> (x -> r) -> r)
Re-bind an unrelated action

-> (a -> r)
return the result

-> Function ps ((q -> r) -> r)
Handle the foremost action

-> Eff ((k >: Action ps q) ': xs) a

-> r

Specialised version of peelEff for Actions. You can pass a function a -> b -> ... -> (q -> r) -> r as a handler for Action '[a, b, ...] q.

peelAction0

Arguments

:: Function ps (Eff xs q)
Handle the foremost action

-> Eff ((k >: Action ps q) ': xs) a

-> Eff xs a

Non continuation-passing variant of peelAction.

transformers-compatible actions and handlers

Reader

type ReaderEff = (:~:)

The reader monad is characterised by a type equality between the result type and the enviroment type.

askEff :: forall k r xs. Associate k (ReaderEff r) xs => Proxy k -> Eff xs r

Fetch the environment.

asksEff :: forall k r xs a. Associate k (ReaderEff r) xs => Proxy k -> (r -> a) -> Eff xs a

Pass the environment to a function.

localEff :: forall k r xs a. Associate k (ReaderEff r) xs => Proxy k -> (r -> r) -> Eff xs a -> Eff xs a

Modify the enviroment locally.

runReaderEff :: forall k r xs a. Eff ((k >: ReaderEff r) ': xs) a -> r -> Eff xs a

Run the frontal reader effect.

State

type State s = StateT s Identity

A state monad parameterized by the type s of the state to carry.

The return function leaves the state unchanged, while >>= uses the final state of the first computation as the initial state of the second.

getEff :: forall k s xs. Associate k (State s) xs => Proxy k -> Eff xs s

Get the current state.

getsEff :: forall k s a xs. Associate k (State s) xs => Proxy k -> (s -> a) -> Eff xs a

Pass the current state to a function.

putEff :: forall k s xs. Associate k (State s) xs => Proxy k -> s -> Eff xs ()

Replace the state with a new value.

modifyEff :: forall k s xs. Associate k (State s) xs => Proxy k -> (s -> s) -> Eff xs ()

Modify the state.

stateEff :: forall k s xs a. Associate k (State s) xs => Proxy k -> (s -> (a, s)) -> Eff xs a

Lift a state modification function.

runStateEff :: forall k s xs a. Eff ((k >: State s) ': xs) a -> s -> Eff xs (a, s)

Run the frontal state effect.

execStateEff :: forall k s xs a. Eff ((k >: State s) ': xs) a -> s -> Eff xs s

Run the frontal state effect and return the final state.

evalStateEff :: forall k s xs a. Eff ((k >: State s) ': xs) a -> s -> Eff xs a

Run the frontal state effect and return the final result.

Writer

type WriterEff w = (,) w

(,) already is a writer monad.

writerEff :: forall k w xs a. Associate k (WriterEff w) xs => Proxy k -> (a, w) -> Eff xs a

Write the second element and return the first element.

tellEff :: forall k w xs. Associate k (WriterEff w) xs => Proxy k -> w -> Eff xs ()

Write a value.

listenEff :: forall k w xs a. (Associate k (WriterEff w) xs, Monoid w) => Proxy k -> Eff xs a -> Eff xs (a, w)

Squash the outputs into one step and return it.

passEff :: forall k w xs a. (Associate k (WriterEff w) xs, Monoid w) => Proxy k -> Eff xs (a, w -> w) -> Eff xs a

Modify the output using the function in the result.

runWriterEff :: forall k w xs a. Monoid w => Eff ((k >: WriterEff w) ': xs) a -> Eff xs (a, w)

Run the frontal writer effect.

execWriterEff :: forall k w xs a. Monoid w => Eff ((k >: WriterEff w) ': xs) a -> Eff xs w

Run the frontal state effect.

Maybe

type MaybeEff = Const ()

An effect with no result

runMaybeEff :: forall k xs a. Eff ((k >: MaybeEff) ': xs) a -> Eff xs (Maybe a)

Run an effect which may fail in the name of k.

Either

type EitherEff = Const

Throwing an exception

throwEff :: Associate k (EitherEff e) xs => Proxy k -> e -> Eff xs a

Throw an exception e, throwing the rest of the computation away.

catchEff :: forall k e xs a. Associate k (EitherEff e) xs => Proxy k -> Eff xs a -> (e -> Eff xs a) -> Eff xs a

Attach a handler for an exception.

runEitherEff :: forall k e xs a. Eff ((k >: EitherEff e) ': xs) a -> Eff xs (Either e a)

Run an action and abort on throwEff.

Iter

data Identity a :: * -> *

Identity functor and monad. (a non-strict monad)

Since: 4.8.0.0

tickEff :: Associate k Identity xs => Proxy k -> Eff xs ()

Put a milestone on a computation.

runIterEff :: Eff ((k >: Identity) ': xs) a -> Eff xs (Either a (Eff ((k >: Identity) ': xs) a))

Run a computation until the first call of tickEff.

Cont

data ContT k r m a :: forall k. k -> (k -> *) -> * -> *

The continuation monad transformer. Can be used to add continuation handling to any type constructor: the Monad instance and most of the operations do not require m to be a monad.

ContT is not a functor on the category of monads, and many operations cannot be lifted through it.

contEff :: Associate k (ContT r m) xs => Proxy k -> ((a -> m r) -> m r) -> Eff xs a

Place a continuation-passing action.

runContEff :: forall k r xs a. Eff ((k >: ContT r (Eff xs)) ': xs) a -> (a -> Eff xs r) -> Eff xs r

Unwrap a continuation.