Data.Extensible.Field

Description

Flexible records and variants

Documentation

newtype Field h kv

A Field h (k ':> v) is h v annotated with the field name k.

Field :: (v -> *) -> Assoc k v -> *

Constructors

Field

Fields

getField :: h (AssocValue kv)

(@=) :: Wrapper h => FieldName k -> Repr h v -> Field h (k :> v) infix 1

Annotate a value by the field name.

foo :: Record '["num" >: Int, "str" >: String]
foo = #num = 42
  <: #str = "foo"
  <: nil

(<@=>) :: (Functor f, Wrapper h) => FieldName k -> f (Repr h v) -> Comp f (Field h) (k :> v) infix 1

Lifted (@=) foo :: IO (Record '["num" >: Int, "str" >: String]) foo = hsequence $ #num <=> readLn #str <@= getLine <: nil @

(@:>) :: FieldName k -> h v -> Field h (k :> v) infix 1

Annotate a value by the field name without Wrapper.

(@==) :: FieldName (k :: Symbol) -> v -> Field Identity (k :> v) infix 1

Kind-monomorphic, unwrapped version of ('=')@

type FieldOptic k = forall kind. forall f p t xs h v. (Extensible f p t, Associate k v xs, Labelling k p, Wrapper h) => Optic' p f (t (Field h) xs) (Repr h v)

FieldOptic s is a type of optics that points a field/constructor named s.

The yielding fields can be Lenses for Records and Prisms for Variants.

FieldOptic "foo" = Associate "foo" a xs => Lens' (Record xs) a
FieldOptic "foo" = Associate "foo" a xs => Prism' (Variant xs) a

FieldOptics can be generated using mkField defined in the Data.Extensible.TH module.

type FieldName k = Optic' (LabelPhantom k) Proxy (Inextensible (Field Proxy) '[k :> ()]) ()

When you see this type as an argument, it expects a FieldLens. This type is used to resolve the name of the field internally.

liftField :: (g (AssocValue kv) -> h (AssocValue kv)) -> Field g kv -> Field h kv

Lift a function for the content.

liftField2 :: (f (AssocValue kv) -> g (AssocValue kv) -> h (AssocValue kv)) -> Field f kv -> Field g kv -> Field h kv

Lift a function for the content.

Records and variants

type RecordOf h = (:*) (Field h)

The type of records which contain several fields.

RecordOf :: (v -> *) -> [Assoc k v] -> *

type Record = RecordOf Identity

Simple record

emptyRecord :: Record '[]

An empty Record.

type VariantOf h = (:|) (Field h)

The dual of RecordOf

VariantOf :: (v -> *) -> [Assoc k v] -> *

type Variant = VariantOf Identity

Simple variant

Matching

matchWithField :: (forall x. f x -> g x -> r) -> RecordOf f xs -> VariantOf g xs -> r

Select a corresponding field of a variant.

matchField :: RecordOf (Match h r) xs -> VariantOf h xs -> r

Pattern matching on a Variant

Key / value

type family AssocKey (kv :: Assoc k v) :: k where ...

Take the type of the key

Equations

AssocKey (k :> v) = k

type family AssocValue (kv :: Assoc k v) :: v where ...

Take the type of the value

Equations

AssocValue (k :> v) = v

class (pk (AssocKey kv), pv (AssocValue kv)) => KeyValue pk pv kv

Combined constraint for Assoc

proxyAssocKey :: proxy kv -> Proxy (AssocKey kv)

Proxy-level AssocKey. This is useful when using symbolVal.

Internal

data LabelPhantom s a b

A ghostly type which spells the field name

type family Labelling s p :: Constraint where ...

Signifies a field name internally

Equations

Labelling s (LabelPhantom t) = s ~ t

Labelling s p = ()

data Inextensible h xs

The trivial inextensible data type