Data.Extensible.Internal

Description

A bunch of combinators that contains magic

Membership

data Membership xs x

The position of x in the type level set xs.

getMemberId :: Membership xs x -> Int

mkMembership :: Int -> Q Exp

Generates a Membership that corresponds to the given ordinal (0-origin).

reifyMembership :: Int -> (forall x. Membership xs x -> r) -> r

leadership :: Membership (y ': xs) x -> ((x :~: y) -> r) -> (Membership xs x -> r) -> r

Embodies a type equivalence to ensure that the Membership points the first element.

compareMembership :: Membership xs x -> Membership xs y -> Either Ordering (x :~: y)

Compare two Memberships.

impossibleMembership :: Membership '[] x -> r

There is no Membership of an empty list.

here :: Membership (x ': xs) x

The Membership points the first element

navNext :: Membership xs y -> Membership (x ': xs) y

The next membership

Member class

class Member xs x where

x is a member of xs

Minimal complete definition

membership

Methods

membership :: Membership xs x

remember :: forall xs x r. Membership xs x -> (Member xs x => r) -> r

Remember that Member xs x from Membership.

type (∈) x xs = Member xs x

Unicode flipped alias for Member

type family FindType (x :: k) (xs :: [k]) :: [Nat] where ...

FindType types

Equations

FindType x (x ': xs) = Zero ': FindType x xs

FindType x (y ': ys) = MapSucc (FindType x ys)

FindType x '[] = '[]

Association

data Assoc k v

The kind of key-value pairs

Constructors

k :> v infix 0

type (>:) = (:>)

A synonym for (:>)

class Associate k v xs | k xs -> v where

Associate k v xs is essentially identical to (k :> v) ∈ xs , but the type v is inferred from k and xs.

Minimal complete definition

association

Methods

association :: Membership xs (k :> v)

type family FindAssoc (key :: k) (xs :: [Assoc k v]) where ...

Equations

FindAssoc k ((k :> v) ': xs) = (Zero :> v) ': MapSuccKey (FindAssoc k xs)

FindAssoc k ((k' :> v) ': xs) = MapSuccKey (FindAssoc k xs)

FindAssoc k '[] = '[]

Sugar

type family Elaborate (key :: k) (xs :: [v]) :: Elaborated k v where ...

Equations

Elaborate k '[] = Missing k

Elaborate k '[x] = Expecting x

Elaborate k xs = Duplicate k

data Elaborated k v

A readable type search result

Constructors

Expecting v

Missing k

Duplicate k

Miscellaneous

data Nat

Type level binary number

Constructors

Zero

DNat Nat

SDNat Nat

class KnownPosition n where

Converts type naturals into Word.

Minimal complete definition

theInt

Methods

theInt :: proxy n -> Int

type family Succ (x :: Nat) :: Nat where ...

The successor of the number

Equations

Succ Zero = SDNat Zero

Succ (DNat n) = SDNat n

Succ (SDNat n) = DNat (Succ n)

type family Head (xs :: [k]) :: k where ...

Equations

Head (x ': xs) = x

type family Last (x :: [k]) :: k where ...

Equations

Last '[x] = x

Last (x ': xs) = Last xs

module Data.Type.Equality

module Data.Proxy