Data.Extensible.Match

Description

Pattern matching

Documentation

matchWith :: (forall x. f x -> g x -> r) -> (f :* xs) -> (g :| xs) -> r

Retrieve the contents so that they matches and pass both to the given function.

newtype Match h r x

Turn a wrapper type into a clause for it.

Constructors

Match

Fields

runMatch :: h x -> r

match :: (Match h a :* xs) -> (h :| xs) -> a

O(1) Perform pattern matching.

mapMatch :: (a -> b) -> Match h a x -> Match h b x

Applies a function to the result of Match.

caseOf :: (h :| xs) -> (Match h a :* xs) -> a infix 0

Flipped match