Data.Extensible.Struct

Description

Mutable structs

Mutable struct

data Struct s h xs

Mutable type-indexed struct.

set :: PrimMonad m => Struct (PrimState m) h xs -> Membership xs x -> h x -> m ()

Write a value in a Struct.

get :: PrimMonad m => Struct (PrimState m) h xs -> Membership xs x -> m (h x)

Read a value from a Struct.

new :: forall h m xs. (PrimMonad m, Generate xs) => (forall x. Membership xs x -> h x) -> m (Struct (PrimState m) h xs)

Create a new Struct using the supplied initializer.

newRepeat :: forall h m xs. (PrimMonad m, Generate xs) => (forall x. h x) -> m (Struct (PrimState m) h xs)

Create a Struct full of the specified value.

newFor :: forall proxy c h m xs. (PrimMonad m, Forall c xs) => proxy c -> (forall x. c x => Membership xs x -> h x) -> m (Struct (PrimState m) h xs)

Create a new Struct using the supplied initializer with a context.

newFromHList :: forall h m xs. PrimMonad m => HList h xs -> m (Struct (PrimState m) h xs)

Create a new Struct from an HList.

data WrappedPointer s h a where

A pointer to an element in a Struct.

Constructors

WrappedPointer :: !(Struct s h xs) -> !(Membership xs x) -> WrappedPointer s h (Repr h x)

(-$>) :: forall k h xs v s. Associate k v xs => Struct s h xs -> Proxy k -> WrappedPointer s h (Repr h (k :> v))

Get a WrappedPointer from a name.

Atomic operations

atomicModify :: PrimMonad m => Struct (PrimState m) h xs -> Membership xs x -> (h x -> (h x, a)) -> m a

atomicModify' :: PrimMonad m => Struct (PrimState m) h xs -> Membership xs x -> (h x -> (h x, a)) -> m a

atomicModify_ :: PrimMonad m => Struct (PrimState m) h xs -> Membership xs x -> (h x -> h x) -> m (h x)

atomicModify'_ :: PrimMonad m => Struct (PrimState m) h xs -> Membership xs x -> (h x -> h x) -> m (h x)

Immutable product

data h :* s

The type of extensible products.

(:*) :: (k -> *) -> [k] -> *

unsafeFreeze :: PrimMonad m => Struct (PrimState m) h xs -> m (h :* xs)

Turn Struct into an immutable product. The original Struct may not be used.

newFrom :: forall g h m xs. PrimMonad m => (g :* xs) -> (forall x. Membership xs x -> g x -> h x) -> m (Struct (PrimState m) h xs)

Create a new Struct using the contents of a product.

hlookup :: Membership xs x -> (h :* xs) -> h x

Get an element in a product.

hlength :: (h :* xs) -> Int

The size of a product.

hfoldrWithIndex :: (forall x. Membership xs x -> h x -> r -> r) -> r -> (h :* xs) -> r

Right-associative fold of a product.

thaw :: PrimMonad m => (h :* xs) -> m (Struct (PrimState m) h xs)

Create a new Struct from a product.

hfrozen :: (forall s. ST s (Struct s h xs)) -> h :* xs

Create a product from an ST action which returns a Struct.

hmodify :: (forall s. Struct s h xs -> ST s ()) -> (h :* xs) -> h :* xs

toHList :: forall h xs. (h :* xs) -> HList h xs

Convert a product into an HList.