Data.Extensible.Tangle

Description

Extensible tangles

Documentation

newtype TangleT h xs m a

TangleT h xs m is the monad of computations that may depend on the elements in xs.

Constructors

TangleT

Fields

unTangleT :: RWST (Comp (TangleT h xs m) h :* xs) () (Nullable h :* xs) m a

lasso :: forall k v m h xs. (Monad m, Associate k v xs, Wrapper h) => FieldName k -> TangleT h xs m (Repr h (k :> v))

Hitch an element associated to the FieldName through a wrapper.

hitchAt :: Monad m => Membership xs x -> TangleT h xs m (h x)

Take a value from the tangles. The result is memoized.

runTangleT

Arguments

:: Monad m

=> (Comp (TangleT h xs m) h :* xs)
tangle matrix

-> (Nullable h :* xs)
pre-calculated values

-> TangleT h xs m a

-> m (a, Nullable h :* xs)

Run a TangleT action and return the result and the calculated values.

evalTangleT

Arguments

:: Monad m

=> (Comp (TangleT h xs m) h :* xs)
tangle matrix

-> (Nullable h :* xs)
pre-calculated values

-> TangleT h xs m a

-> m a

Run a TangleT action.

runTangles

Arguments

:: Monad m

=> (Comp (TangleT h xs m) h :* xs)
tangle matrix

-> (Nullable h :* xs)
pre-calculated values

-> m (h :* xs)

Run tangles and collect all the results as a Record.