Data.Extensible.Wrapper

Description

Documentation

class Wrapper h where

The extensible data types should take k -> * as a parameter. This class allows us to take a shortcut for direct representation.

Associated Types

type Repr h (v :: k) :: *

Repr h v is the actual representation of h v.

Methods

_Wrapper :: (Functor f, Profunctor p) => Optic' p f (h v) (Repr h v)

This is an isomorphism between h v and Repr h v.

_Wrapper :: Iso' (h v) (Repr h v)

_WrapperAs :: (Functor f, Profunctor p, Wrapper h) => proxy v -> Optic' p f (h v) (Repr h v)

Restricted version of _Wrapper. It is useful for eliminating ambiguousness.

newtype Const' a x

Poly-kinded Const

Constructors

Const'

Fields

getConst' :: a

newtype Comp f g a

Poly-kinded composition

Constructors

Comp

Fields

getComp :: f (g a)

comp :: Functor f => (a -> g b) -> f a -> Comp f g b

Wrap a result of fmap

data Prod f g a

Poly-kinded product

Constructors

Prod (f a) (g a)